asteroid repairs
i have changed the ai for the asteroids. they all now gravitate towards earth, but at different speeds depending on which spawners they originate from. I also added a maximum amount of asteroids and a cleanup mechanism. The maximum can be surpassed, but it would be difficult to maintain for long.