text colors, asteroid variations, speed increases & mining during orbit
the asteroids now give points while orbitting- as if they are being mined for resources. When you click on an asteroid you get 6 points. also a change on click to destroy, it takes 2 seconds with an animation. there are new asteroids which give more resources. the asteroids fly in faster as your score goes higher